Added pet scaling
CastSpell() checks if the caster is a pet and if the SCALE_PET_SPELLS server property is true.  If so, it scales spell damage and values linearly with a level 50 value based on PC spells.  The SCALE_PET_SPELLS_PERCENTAGE property is then used to set the spell to a percentage of the player's equivalent spell at that level.

This will make pets weaker in the early game but more powerful end game.  For example, a level 7 enchanter cast pet has a 100 DD spell, which is incredible at that level, but still has that same 100 DD spell at level 50.  If enabled via server properties, pets will be much more consistently powerful throughout the range of available levels.