Removed hardcoded efficiency values (#86)
* Removed hardcoded efficiency values

Removed hardcoded efficiency values in functions that should have been using the Efficiency accessor instead.  Modified multiplier to effectiveness to not assume base effectiveness is 1.

Also removed special characters that always pop up as a change so that doesn't happen in future.

* Replaced hardcoded effectiveness in AttackDamage()

ontheDOL found that GamePlayer already factors in effectiveness in it's overloaded functions, so changing it to use Effectiveness would lead to player's damage going down to (0.7*0.7) rather than just 0.7.

OnTick() change impacts pet melee and ranged damage, and stays in.