Fixed LootGenerator exceptions adding first item (#69)
Changed to create loot using an existing ItemTemplate from the DB. Creating a new one causes the server to try to add the ItemTemplate to the ItemTemplate table, causing an exception which then mucks up the player's inventory.  This only happens if there isn't already an item of that type in the player's inventory, otherwise it just increases the count instead.

Also changed LootGeneratorAurulite to be in the DOL.GS rather than DOL.GS.Scripts namespace.