Reduced drop count & considers target
The previous implementation appeared to be a copy/paste of the loot generator for Aurulite or Atlantean Glass.  As a result, it was dropping large stacks of dragon scales when it shouldn't be.

This redo checks for named dragons and drops the correct number of scales for named dragons while still allowing amounts dropped to be determined by server properties, so it can be used for epic and other named dragons rather than using loot templates.