Fixed hangabout theurgist pets (#104)
* Pets die when target dies

On the next Think() after a theurgist pet's target dies, it dies as well rather than standing around until its duration runs out.

* Replaced deep copy of m_target

It's possible a deep copy is needed should something else kill the target, leaving the shallow copy as null to mess up CheckSpells() or StartAttack().  Better to be paranoid.

* Pets die when player dies

Added a check for the owner's status as a trigger for elemental suicide.  Otherwise, they stick around after the owner was killed.

* Removed pet dying when owner dies