autoselect caster for realm spells
This adds caster autoselection for beneficial spells.  Let's say your valkyrie is soloing a mob, and you need to cast an insta-heal.  In live implementation, you have to target yourself, cast the heal, then target the mob.  With this change, the heal spell will automatically target the caster when the spell isn't appropriate for the current target, so you can just heal yourself without having to change targets.

While this isn't a live feature, it feels like a no brainer to me.  Other MMOs from the same generation included this as a feature, and it's ridiculous that live doesn't.  That said, it's really easy to add a server property to enable this if the community wishes.