Changed: Uses dictionaries rather than hashtables
This should be both safer and faster by enforcing typing and skipping a ton of boxing/unboxing.  String/Int type mismatches, for example, will now cause a compiler error rather than allowing an object to be inserted with one type and looked for with another.

Also added members to see if a keep already has characters added to it, and an accessor to return Height to make it more intuitive.