Fixed issue with archer mobs at max range (#117)
* Fixed issue with archer mobs at max range

Sometimes, mobs will try to fire arrows but due to some gimmick in the math somewhere be just a hair out of range.  This leads to them starting to fire over and over again without actually firing.  It appears to add multiple ranged fire timers, so the moment the target is in range, all the timers go off at once and do a ridiculous amount of burst damage.

By forcing mobs to move just a tiny bit closer before firing, this issue appears to be resolved.

* Fixed issue with archer mobs at max range 

A 10 offset wasn't enough to stop the problem from happening, and 30 seems to be about as low as I can set it to prevent it.