Server props for keep guard stats (#125)
* Exposed min and max stick range constants

These should be exposed, as follow can be called outside of GameNPC.

Also changed a few readonly statics to consts.

* Reverting back to master

Accidentally uploaded a file to my master branch.  Reverting changes

* Change: ServerProps for keep lords and guards

Stats for keep guards and lords were hardcoded in TemplateManager.  This replaces that with server properties.

Default values are based on what the original hardcoded values were.

* Change: Calculates stats based on serverprops

This was hardcoded, so I'm replacing all that with server properties.

* Change: ServerProps for keep lord/guard stats

Stats for keep guards and lords were hardcoded in TemplateManager.  This replaces that with server properties.

Default values are based on what the original hardcoded values were.

* Change: Added 10x multiplier to STR_MULTIPLIER