Added allow_train_anywhere functionality
I really like Phoenix letting people use the /train command to train anywhere in the world rather than having to run back to their trainer, so I added that functionality.  Since it's not livelike, it's disabled by default and has to be enabled by the allow_train_anywhere server property.