Necro pet spellcasting can be interrupted
Necro pets weren't being interrupted when spellcasting.  I added code to ensure the pet can actually cast spells before adding them to the queue, and used OnAttackedByEnemy() to check for interrupts and cancel spellcasting/clear the spell queue.

That meant that instant spells could be queued behind ones with cast times and be lost if the former spell was interrupted and the queue cleared, so I had instant spells bypass the queue and cast immediately.