Added BAF_MOBS_ATTACK_RANDOM_PLAYERS, expanded to check battlegroups
Live almost always sends BAF mobs after the player who pulled, so I added a server property to choose whether mobs go after random players or after the player who pulled them.

I also expanded the code to look for battlegroups so that it counts and sends mobs after all the players in a raid rather than just the players in the pulling group.  Since BGs can be up to 64 players by default, I tried to make the code as efficient as possible.