Added BAF_MOBS_ATTACK_RANDOM_PLAYERS server property, expanded to check battlegroups (#182)
* Added BAF_MOBS_ATTACK_RANDOM_PLAYERS

Defaults false.

If true, mobs brought by a BAF attack random players in the puller's battlegroup or group.

If false, additional mobs aggro on the puller as on live.

* Added BAF_MOBS_ATTACK_RANDOM_PLAYERS, expanded to check battlegroups

Live almost always sends BAF mobs after the player who pulled, so I added a server property to choose whether mobs go after random players or after the player who pulled them.

I also expanded the code to look for battlegroups so that it counts and sends mobs after all the players in a raid rather than just the players in the pulling group.  Since BGs can be up to 64 players by default, I tried to make the code as efficient as possible.

* Fixed GetPlayersInTheBattleGroup()

GetPlayersInTheBattleGroup() was checking the wrong enumerator and always returning an empty array.

The new implementation returns an array of players, and should be a little more efficient as it uses an ArrayList sized to the current number of members in the BG rather than the maximum possible number of members.