Improved pet spell scaling (#184)
* Added pet spell scaling to style effects

* Make SortSpells() and SortStyles() public

* Added ScalePetSpell()

Added a method to scale a pet spell, so it's easily accessible in spell handlers and the style processor

* Removed pet spell scaling

Pet scaling no longer needs to be added to specific spell handlers

* Added pet spell scaling

Pet spells are now scaled when the pet is first created, rather than scaling them on each cast.

* Redid pet spell scaling

Code to scale pet spells is not in a method on the pet class.  Spells are scaled as they're added in SummonSpellHandler, but subspells still need to be scaled in SpellHandler.