Tickets are now automatically given to merchant when bought
On live, when you buy a ticket, it's immediately given to the merchant to start the trip.  I overloaded OnPlayerBuy() to add that feature.

It checks for "ticket to" in the ticket name, but also checks for "ticket" as the description.  This should make it function for English servers immediately, while giving non-English servers a way to make it work.