Added BD commander strings
Added more BD commander options to support features which are now implemented.  This file is necessary for BD pets to work, so I created copies in other language directories so at least the options will be available, albeit in the wrong language.

German was the only language with most options defined, so I tried to use it as much as possible.

Strings now use single word language entries whenever possible, so if something has to change it only has to be changed in one place.