Set mob_2h_bonus_damage false by default
Paradoxically, there's only a handful of mobs in Eve that can have a 2H weapon and a block chance, and they only use 2H weapons.  An exception is the Angelic Visage, which uses a 2H weapon and has a 20% block chance.

Since setting this true would be an arbitrary 5% damage bonus on some mobs that can't ever block anyway, I'm having 2H bonus damage default to false for mobs.