Added support for PetSummonThreshold > 100%
Also:
Added PetResummonTime accessor.
Changed SummonThreshold to PetSummonThreshold to match other accessors.
Changed to override TakeDamage() to move a lot of check out of Health.set
Removed torch flag from code; it should be done in the DB anyway.

This still needs a little work for PetSummonThreshold > 100%, and likely an entire new GamePet variation is needed for NPC pets.