Fix Quest Mob Selection (#244)
Some quest mob names are reused (Elizabeth), and others are duplicated in strange places in Eve.  This made some quests impossible.  Quests now check for name, region, and x/y coordinates to make sure the correct NPC is being selected rather than assuming the first one in the list is the one we want.