Improved Code to Enable Craft Button
Old code enabled the craft button for items if the player's primary crafting skill was alchemy or spellcrafting.  This means that the craft button could be enabled for items that weren't player crafted, and that servers which give access to all crafting skills never enable the crafting button.

New code checks if the item is player crafted, is from one of the crafting skills which can be imbued or tinctured, and the player has either alchemy or spellcrafting skills.