Now handles crafted seals, works at max realm rank
More valuable seals created via crafting are given their proper values rather than all seals being treated equally and giving somebody turning in an effulgent dreaded seal 1/250th of it's value.

The collector now takes seals and grants BPs when the player is at max realm rank rather than refusing to take the seal entirely.

There was a fair bit of clean up as well.  The script this was based on had several people fiddle with it before it got to me, and there was a lot of room for improvement.  It's possible I'm better at this than I was three years ago.