Added CheckStealth() Method
Added CheckStealth() Method, and stuck it in the end of Think().

This allows NPCs to visibly unstealth when attacking and restealth when out of combat, but also fixes problems with NPC brains checking IsStealthed assuming players, and as a result not being able to attack or defend themselves against stealthed NPCs.