Added Stealth Accessors and Methods
Added a CanStealth accessor to indicate if an NPC is cable of stealthing, and a Stealth() method akin to the one used by player classes.

When Flags is updated, it sends a packet to update the model on the client side, so NPCs stealth and unstealth visually as well as logically.

Also added accessors for the other possible eFlags, because we have tons of bitmasks everywhere and I was getting annoyed constantly retyping them.  :-)