Added CanStealth and ControlledBody, removed StealthPets()
Moved CanStealth from GameNPC to GameLiving so a common accessor can be used for GameNPC and GamePlayer.

ControlledBody returns the first pet's body, with proper null checks, to allow easier and more intuitive pattern matching.

StealthPets() was intended for Blanket of Camouflage, and wasn't a viable approach.

Added Skills region.