Reworked Blanket of Camouflage
Testing revealed a bunch of problems with Blanket of Camouflage, which have now been addressed.

It now properly blankets pets and subpets as well as players.  Pets have event handlers to remove camouflage, and remove camoflouge from their owners when they lose it themselves.

Storing the spell effect in a local variable lead to the local variable not always matching the event sender, leading to things like a player moving remaining camouflaged while removing camouflage from another target.  Effects are now pulled from targets instead.

Choosing valid targets is now done in SelectTargets() and ApplyEffectOnTarget(), as OnEffectStart() happens after the effect has been added, but without the event handlers to remove it.  Stealthers don't get the effect at all, but are stealthed.  This appears to be a problem with other ML abilities, but that will be addressed in future PRs.

There was a completely unnecessary self target check, which was removed.

Necros remove camouflage from their pet when they lose it themselves, so they can't move around or cast spells through their pet without exposing it to attack.