Create a GameInventoryItem when interacting with a GameInventoryObject
Every item added to the player's inventory needs to be a GameInventoryItem type.
There exists a condition in GameInventoryObject's which allow to exchange two items
from a vault, during which the item received loses its type and uses InventoryItem
instead.
A player could exploit this behaviour to prevent items from losing condition (for
example).