Allow NPC Protect Effects
ProtectEffect no longer assumes both the source and target are players, allowing the effect to work and no longer throw exceptions when the effect is being applied by and/or to NPCs.

Removed a superfluous player group check, as that is already done in ProtectAbilityHandler.Execute().

I'm specifically using this for custom pets that protect their owner, but it could in theory be used in other contexts.  Somebody could create a group vs group fight between NPCs, and have tank NPCs apply protect to healing NPCs, for example.

While I was there, I cleaned up the code and addressed compiler warnings.