Fix AddToAggroList() LOS
AddToAggroList() was sending a LOS request to the client, and expecting an instantaneous response stored in m_AggroLOS.  Not only does that not get the actual output until the next Think() cycle, but it also risks applying the result to the wrong target.  I've split AddtoAggroList() into two parts to allow for a proper callback response akin to StartAttack() and ContinueStartAttack().  Aggroamounts are stored in a list by ObjectID while the LOS check is pending, and I've tried to streamline that as much as possible.

I've also added a server property to determine whether or not to check LOS when adding aggro to a keep guard, to replace them being hardcoded to not check.  The property defaults to false, so keep guard behaviour shouldn't change.  I'm seeing some odd behaviour by keep guards when set to true, but I want to make sure the LOS code in general is good before I get into the weeds puzzling that out.

Mobs checking for LOS will be impacted by this, but my testing hasn't revealed any unusual behaviour.  It should, if anything, be more efficient by adding aggro before the next Think() cycle and saving redundant CheckNPC() and CheckPlayer() calls and their associated LOS checks.